api/requirements.txt
fastapi
gunicorn
uvicorn
openrouteservice
SQLAlchemy
psycopg2-binary
passlib
pydantic
pytest
pandas
kagglehub
pytest-asyncio
pytest-mock
python-dateutil
python-dotenv
python-editor
python-jose
python-multipart
typing_extensions
redis
email-validator
shapely
geopandas
numpy
httpx
pathlib
argon2_cffi
kagglehub